[cleanup] Split table synchronizer to avoid tangled method.
