+++ plugins/org.eclipse.sirius.table/src/org/eclipse/sirius/table/business/internal/dialect/TableDialectServices.java
@@ -37,14 +37,17 @@ import org.eclipse.sirius.common.tools.api.util.StringUtil;
 import org.eclipse.sirius.ecore.extender.business.api.accessor.ModelAccessor;
 import org.eclipse.sirius.table.business.api.refresh.DTableSynchronizer;
 import org.eclipse.sirius.table.business.internal.dialect.description.TableInterpretedExpressionQuery;
+import org.eclipse.sirius.table.business.internal.refresh.CrossTableSynchronizer;
 import org.eclipse.sirius.table.business.internal.refresh.DTableElementSynchronizer;
-import org.eclipse.sirius.table.business.internal.refresh.DTableSynchronizerImpl;
+import org.eclipse.sirius.table.business.internal.refresh.EditionTableSynchronizer;
 import org.eclipse.sirius.table.metamodel.table.DCell;
 import org.eclipse.sirius.table.metamodel.table.DColumn;
 import org.eclipse.sirius.table.metamodel.table.DLine;
 import org.eclipse.sirius.table.metamodel.table.DTable;
 import org.eclipse.sirius.table.metamodel.table.DTableElement;
 import org.eclipse.sirius.table.metamodel.table.TableFactory;
+import org.eclipse.sirius.table.metamodel.table.description.CrossTableDescription;
+import org.eclipse.sirius.table.metamodel.table.description.EditionTableDescription;
 import org.eclipse.sirius.table.metamodel.table.description.TableDescription;
 import org.eclipse.sirius.table.tools.internal.Messages;
 import org.eclipse.sirius.table.tools.internal.command.TableCommandFactory;
@@ -128,15 +131,39 @@ public class TableDialectServices extends AbstractRepresentationDialectServices 
             DTable table = (DTable) representation;
             IInterpreter interpreter = SiriusPlugin.getDefault().getInterpreterRegistry().getInterpreter(table.getTarget());
             ModelAccessor accessor = SiriusPlugin.getDefault().getModelAccessorRegistry().getModelAccessor(representation);
-            TableDescription description = table.getDescription();
-            DTableSynchronizer sync = new DTableSynchronizerImpl(description, createElementSynchronizer(accessor, interpreter));
-            sync.setTable(table);
-            sync.refresh(new SubProgressMonitor(monitor, 1));
+
+            DTableSynchronizer sync = createTableSynchronizer(table.getDescription(), accessor, interpreter);
+            if (sync != null) {
+                sync.setTable(table);
+                sync.refresh(new SubProgressMonitor(monitor, 1));
+            }
         } finally {
             monitor.done();
         }
     }
 
+    /**
+     * Creates a synchronizer for the provide table.
+     * 
+     * @param description of table to synchronize
+     * @param accessor of model
+     * @param interpreter of expressions
+     * @return synchronizer
+     */
+    public DTableSynchronizer createTableSynchronizer(TableDescription description, final ModelAccessor accessor, final IInterpreter interpreter) {
+        DTableElementSynchronizer elementSyn = createElementSynchronizer(accessor, interpreter);
+        DTableSynchronizer result;
+        if (description instanceof CrossTableDescription) {
+            result = new CrossTableSynchronizer((CrossTableDescription) description, elementSyn);
+        } else if (description instanceof EditionTableDescription) {
+            result = new EditionTableSynchronizer((EditionTableDescription) description, elementSyn);
+        } else {
+            result = null;
+        }
+        return result;
+    }
+    
+    
     private DTableElementSynchronizer createElementSynchronizer(final ModelAccessor accessor, final IInterpreter interpreter) {
         return new DTableElementSynchronizer(accessor, interpreter);
     }
