+++ plugins/org.eclipse.sirius.tests.junit/src/org/eclipse/sirius/tests/unit/table/tests/srs/REQ_157_Test.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2010, 2014 THALES GLOBAL SERVICES.
+ * Copyright (c) 2010, 2023 THALES GLOBAL SERVICES.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -18,7 +18,6 @@ import java.util.List;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import org.eclipse.sirius.table.business.api.refresh.DTableSynchronizer;
-import org.eclipse.sirius.table.business.internal.refresh.DTableSynchronizerImpl;
 import org.eclipse.sirius.table.metamodel.table.DTable;
 import org.eclipse.sirius.table.metamodel.table.TableFactory;
 import org.eclipse.sirius.table.metamodel.table.description.TableDescription;
@@ -38,7 +37,7 @@ public class REQ_157_Test extends TableTestCase {
         final TableDescription desc = find("Colored Classes Table");
         assertNotNull("Unit test data is not correct", desc);
 
-        final DTableSynchronizer sync = new DTableSynchronizerImpl(desc, accessor, interpreter);
+        final DTableSynchronizer sync = createTableSynchronizer(desc);
 
         final DTable newTable = TableFactory.eINSTANCE.createDTable();
         newTable.setDescription(desc);
@@ -68,7 +67,7 @@ public class REQ_157_Test extends TableTestCase {
         final TableDescription desc = find("Colored Classes Table");
         assertNotNull("Unit test data is not correct", desc);
 
-        final DTableSynchronizer sync = new DTableSynchronizerImpl(desc, accessor, interpreter);
+        final DTableSynchronizer sync = createTableSynchronizer(desc);
 
         final DTable newTable = TableFactory.eINSTANCE.createDTable();
         newTable.setDescription(desc);
@@ -102,7 +101,7 @@ public class REQ_157_Test extends TableTestCase {
         final TableDescription desc = find("Colored Classes Table");
         assertNotNull("Unit test data is not correct", desc);
 
-        final DTableSynchronizer sync = new DTableSynchronizerImpl(desc, accessor, interpreter);
+        final DTableSynchronizer sync = createTableSynchronizer(desc);
 
         final DTable newTable = TableFactory.eINSTANCE.createDTable();
         newTable.setDescription(desc);
